+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - new(db): Renamed and transformed "entries" into "places"
 - new(db): Added status log for place revisions
 - fix(db): Removed categories from database
+- fix(db): Fixed invalid event dates and validate all new dates
 
 ## v0.7.2 (2019-11-08)